requirements.txt
google-cloud-resource-manager==1.17.0
google-cloud-run==0.10.1
google-cloud-functions==1.13.0